# docs/swagger.yaml
basePath: /
...
info:
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Timescale Wrapper API
  version: "0.1"
paths:
...
  /download:
    get:
      consumes:
      - application/json
      description: download CSV
      parameters:
      - description: JSON encoded QueriesRequestElement
        in: query
        name: query
        required: true
        type: string
      - description: 'Textual representation of the date ''Mon Jan 2 15:04:05 -0700
          MST 2006''. Example: 2006-01-02T15:04:05.000Z07:00 would format timestamps
          as rfc3339 with ms precision. Find details here: https://golang.org/pkg/time/#Time.Format'
        in: query
        name: time_format
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: download
  /download/{secret}:
    get:
      description: downloads CSV file with previously prepared secret
      parameters:
      - description: secret
        in: path
        name: secret
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: download
...
  /prepare-download:
    get:
      consumes:
      - application/json
      description: genartes a secret for later download. can be used in native browser
        downloads
      parameters:
      - description: JSON encoded QueriesRequestElement
        in: query
        name: query
        required: true
        type: string
      - description: 'Textual representation of the date ''Mon Jan 2 15:04:05 -0700
          MST 2006''. Example: 2006-01-02T15:04:05.000Z07:00 would format timestamps
          as rfc3339 with ms precision. Find details here: https://golang.org/pkg/time/#Time.Format'
        in: query
        name: time_format
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: prepare download
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"